requirements.txt
requests
beautifulsoup4
python-dotenv
Pillow
atproto
tweepy
gspread
twilio
pyyaml
anthropic
edge-tts
google-api-python-client
google-auth-oauthlib
